chore: move specific utils closer to relevant code
